Extract driver unit tests
